slack/requirements.txt
streamlit==1.29.0
anthropic==0.7.0
python-dotenv==1.0.0
pandas==2.1.3
pydantic==2.5.2
docx2pdf==0.1.8
python-docx==1.0.1